requirements.txt
# Core Dependencies
langgraph>=0.0.30
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# LLM & Embeddings
openai>=1.0.0
anthropic>=0.8.0
sentence-transformers>=2.2.0
# LLM APIs
anthropic>=0.39.0

# LLM APIs - Google Gemini (FREE!)
google-generativeai>=0.8.0
# Vector Database
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# NER & NLP
spacy>=3.7.0
stanza>=1.7.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Monitoring & Logging
loguru>=0.7.0